# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AprovaFácil API",
    "description": "API para sistema de estudos e preparação para concursos",
    "version": "1.1.0",
    "contact": {
      "name": "AprovaFacil Team",
      "email": "contato@aprovafacil.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.aprovafacil.com",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
...
    "/mapa-assuntos": {
      "get": {
        "operationId": "getMapaAssuntos",
        "summary": "Listar mapa de assuntos",
        "description": "Retorna o mapa de assuntos organizados por disciplina",
        "tags": [
          "Mapa de Assuntos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "concurso",
            "in": "query",
            "description": "Filtrar por concurso",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "disciplina",
            "in": "query",
            "description": "Filtrar por disciplina específica",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mapa de assuntos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mapa": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "disciplina": {
                            "type": "string",
                            "example": "Direito Constitucional"
                          },
                          "totalAssuntos": {
                            "type": "integer",
                            "example": 25
                          },
                          "assuntosEstudados": {
                            "type": "integer",
                            "example": 15
                          },
                          "progresso": {
                            "type": "number",
                            "example": 0.6
                          },
                          "assuntos": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "example": "assunto-1"
                                },
                                "nome": {
                                  "type": "string",
                                  "example": "Princípios Fundamentais"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Fundamentos da República"
                                },
                                "nivel": {
                                  "type": "string",
                                  "example": "basico"
                                },
                                "status": {
                                  "type": "string",
                                  "example": "estudado"
                                },
                                "questoes": {
                                  "type": "integer",
                                  "example": 10
                                },
                                "acertos": {
                                  "type": "integer",
                                  "example": 8
                                },
                                "taxaAcerto": {
                                  "type": "number",
                                  "example": 0.8
                                },
                                "tempoEstudo": {
                                  "type": "integer",
                                  "example": 3600
                                },
                                "subassuntos": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "id": {
                                        "type": "string",
                                        "example": "subassunto-1"
                                      },
                                      "nome": {
                                        "type": "string",
                                        "example": "Dignidade da Pessoa Humana"
                                      },
                                      "status": {
                                        "type": "string",
                                        "example": "estudado"
                                      },
                                      "progresso": {
                                        "type": "number",
                                        "example": 1
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/mapa-assuntos/{id}": {
      "get": {
        "operationId": "getMapaAssuntosById",
        "summary": "Buscar assunto por ID",
        "description": "Retorna detalhes completos de um assunto específico",
        "tags": [
          "Mapa de Assuntos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do assunto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do assunto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "assunto": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "assunto-1"
                        },
                        "nome": {
                          "type": "string",
                          "example": "Princípios Fundamentais"
                        },
                        "description": {
                          "type": "string",
                          "example": "Fundamentos da República"
                        },
                        "disciplina": {
                          "type": "string",
                          "example": "Direito Constitucional"
                        },
                        "nivel": {
                          "type": "string",
                          "example": "basico"
                        },
                        "status": {
                          "type": "string",
                          "example": "estudado"
                        },
                        "questoes": {
                          "type": "integer",
                          "example": 10
                        },
                        "acertos": {
                          "type": "integer",
                          "example": 8
                        },
                        "taxaAcerto": {
                          "type": "number",
                          "example": 0.8
                        },
                        "tempoEstudo": {
                          "type": "integer",
                          "example": 3600
                        },
                        "dataInicio": {
                          "type": "string",
                          "format": "date"
                        },
                        "dataConclusao": {
                          "type": "string",
                          "format": "date"
                        },
                        "subassuntos": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "example": "subassunto-1"
                              },
                              "nome": {
                                "type": "string",
                                "example": "Dignidade da Pessoa Humana"
                              },
                              "status": {
                                "type": "string",
                                "example": "estudado"
                              },
                              "progresso": {
                                "type": "number",
                                "example": 1
                              },
                              "questoes": {
                                "type": "integer",
                                "example": 5
                              },
                              "acertos": {
                                "type": "integer",
                                "example": 4
                              }
                            }
                          }
                        },
                        "recursos": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tipo": {
                                "type": "string",
                                "example": "apostila"
                              },
                              "titulo": {
                                "type": "string",
                                "example": "Apostila de Direito Constitucional"
                              },
                              "url": {
                                "type": "string",
                                "example": "/apostilas/direito-constitucional"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Assunto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "putMapaAssuntosById",
        "summary": "Atualizar status do assunto",
        "description": "Atualiza o status de estudo de um assunto",
        "tags": [
          "Mapa de Assuntos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do assunto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "nao_estudado",
                      "em_andamento",
                      "estudado",
                      "revisao"
                    ],
                    "description": "Novo status do assunto"
                  },
                  "progresso": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Progresso do estudo (0-1)"
                  },
                  "tempoEstudo": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Tempo adicional de estudo em segundos"
                  }
                }
              },
              "examples": {
                "atualizacao": {
                  "summary": "Exemplo de atualização de status",
                  "value": {
                    "status": "estudado",
                    "progresso": 1,
                    "tempoEstudo": 1800
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Status atualizado com sucesso"
                    },
                    "assunto": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Assunto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Mensagem de erro"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              },
              "message": {
                "type": "string",
                "example": "Mensagem de erro"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}